% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ssdvignettes-version.R
\name{get_ssdvignettes_version}
\alias{get_ssdvignettes_version}
\title{Get the the version of ssdvignettes}
\usage{
get_ssdvignettes_version()
}
\value{
The version of ssdvignettes
}
\description{
Prints the version of ssdvigenttes you have installed
}
\examples{
get_ssdvignettes_version()
}
